+++ NEWS.md
@@ -1,9 +1,11 @@
-# semhelpinghands 0.1.2.3
+# semhelpinghands 0.1.2.4
 
 - Improve the notes in examples. (0.1.2.1)
 - Fix some bugs. (0.1.2.1)
 - Provide the background information on
   `standardizedSolution_boot_ci()`. (0.1.2.2, 0.1.2.3)
+- Add `se_ratios()` for comparing standard errors by
+  different estimators. (0.1.2.4)
 
 # semhelpinghands 0.1.2.0